refactor api class and errorBoundary
