--- src/components/ErrorBoundary/types.ts
@@ -1,11 +1,0 @@
-import { ReactNode } from 'react';
-
-interface IErrorBoundaryProps {
-  children: ReactNode;
-}
-
-interface IErrorBoundaryState {
-  hasError: boolean;
-}
-
-export type { IErrorBoundaryProps, IErrorBoundaryState };
